+++ cmd_chat/client/i18n/__init__.py
@@ -1,0 +1,5 @@
+"""Internationalization module for CMD CHAT."""
+from cmd_chat.client.i18n.translations import TranslationManager
+
+__all__ = ['TranslationManager']
+